fix: eliminate compiler warnings and fix malloc/free mismatch in Capstone
Suppress C4996 (unsafe CRT) via _CRT_SECURE_NO_WARNINGS in CLI project.
Fix C4244 truncation warnings in CombatShell (x86 IAT repair, data directory
restore), Capstone (time_t, uint64_t casts), CompressionData (VmAddroffset),
studData (OEP assignment), and CodeTool (hex conversion). Replace delete[]
with free() for malloc-allocated buffers in Capstone. Replace strcpy with
strncpy for bounded copy. Remove dead debug variables in CompressionData.

+++ CombatShell/CombatShell.cpp
@@ -507,9 +507,9 @@ void UnCompression()
 	{
 		MyVirtualProtect(pDataDirectory, 0x8, PAGE_READWRITE, &Att_old);
 		if (0 != g_stud.s_DataDirectory[i][0])
-			pDataDirectory->VirtualAddress = g_stud.s_DataDirectory[i][0];
+			pDataDirectory->VirtualAddress = (DWORD)g_stud.s_DataDirectory[i][0];
 		if (0 != g_stud.s_DataDirectory[i][1])
-			pDataDirectory->Size = g_stud.s_DataDirectory[i][1];
+			pDataDirectory->Size = (DWORD)g_stud.s_DataDirectory[i][1];
 		MyVirtualProtect(pDataDirectory, 0x8, Att_old, &Att_old);
 		++pDataDirectory;
 	}
@@ -924,12 +924,12 @@ void RepairTheIAT()
 					continue;
 				}
 				PIMAGE_IMPORT_BY_NAME pName = (PIMAGE_IMPORT_BY_NAME)(thunkValue + dwMoudle);
-				FunAddress = SafeGetProcAddressForIat(hModuledll, pName->Name);
+				FunAddress = (DWORD)SafeGetProcAddressForIat(hModuledll, pName->Name);
 			}
 			else
 			{
 				DWORD dwFunOrdinal = IMAGE_ORDINAL32((DWORD)thunkValue);
-				FunAddress = SafeGetProcAddressForIat(hModuledll, (char*)(ULONG_PTR)dwFunOrdinal);
+				FunAddress = (DWORD)SafeGetProcAddressForIat(hModuledll, (char*)(ULONG_PTR)dwFunOrdinal);
 			}
 #endif
